Remove non-working nrf9160dk_nrf9160 overlay for arduino_uno_click shield
It is not currently possible to define a board-specific overlay for a
shield in a module when that shield is defined elsewhere (e.g. in the
Zephyr tree).

See https://github.com/zephyrproject-rtos/zephyr/issues/71113 for more
details.

If this ever becomes a feature that Zephyr provides, we can bring this
overlay back into the module.

Signed-off-by: Chris Wilson <chris@cgnd.dev>