Add startup & on-off delays for Aludel Elixir power regulators
On the Aludel Elixir board, 3.3V power to some devices is gated behind
a hot-swap chip controlled by the CLICK_QWIIC_PWR_CTRL signal. Without a
power-on delay, the MCU will try to communicate with these devices
before the 3.3V supply is stable. In this case, some devices will not
show up on the I2C bus.

This commit adds a power-on delay to the 3.3V hotswap to ensure that
these devices have enough time to initiailze before the MCU attempts to
communicate with them. In addition, this commit also adds a power-on
delay to the 5V power rail that supplies the mikroBUS sockets.

The turn-on delays are variable based on load capacitance/resistance,
but based on the datasheets, the turn-on delays should be less than 1ms
for both regulators. A value of 2ms was chosen for some additional
safety margin.

Signed-off-by: Chris Wilson <chris@cgnd.dev>